fix(gradle-plugin): use dedicated plugin for JDWP configuration
