+++ gradle-plugin/src/main/java/com/itsaky/androidide/gradle/AndroidIDEGradlePlugin.kt
@@ -16,15 +16,8 @@
  */
 package com.itsaky.androidide.gradle
 
-import com.android.build.api.artifact.SingleArtifact
-import com.android.build.api.variant.AndroidComponentsExtension
-import com.android.build.api.variant.ApplicationAndroidComponentsExtension
-import com.android.build.gradle.internal.tasks.factory.TaskConfigAction
-import com.android.build.gradle.internal.tasks.factory.registerTask
-import com.itsaky.androidide.gradle.tasks.AppDebuggerConfigTask
-import com.itsaky.androidide.gradle.utils.JdwpOptions.Companion.jdwpOptions
-import com.itsaky.androidide.tooling.api.ToolingConfig
-import org.gradle.api.GradleException
+import com.itsaky.androidide.tooling.api.LogSenderConfig.PROPERTY_LOGSENDER_ENABLED
+import com.itsaky.androidide.tooling.api.ToolingConfig.PROP_JDWP_INJECT
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.logging.Logging
@@ -41,80 +34,24 @@ class AndroidIDEGradlePlugin : Plugin<Project> {
         private val logger = Logging.getLogger(AndroidIDEGradlePlugin::class.java)
     }
 
-    @Suppress("UnstableApiUsage")
     override fun apply(target: Project) {
         if (target.isTestEnv) {
             logger.lifecycle("Applying ${javaClass.simpleName} to project '${target.path}'")
         }
 
         target.run {
-            val isLogSenderEnabled = false /*if (hasProperty(PROPERTY_LOGSENDER_ENABLED)) {
-                property(PROPERTY_LOGSENDER_ENABLED).toString().toBoolean()
-            } else {
-                // enabled by default
-                true
-            }*/
+            val isLogSenderEnabled = findProperty(PROPERTY_LOGSENDER_ENABLED) == "true"
+            val isDebuggerEnabled = findProperty(PROP_JDWP_INJECT) == "true"
 
             if (plugins.hasPlugin(APP_PLUGIN)) {
                 if (isLogSenderEnabled) {
-                    logger.info("Trying to apply LogSender plugin to project '${project.path}'")
+                    logger.info("Applying LogSender plugin to project '${project.path}'")
                     pluginManager.apply(LogSenderPlugin::class.java)
-                } else {
-                    logger.warn("LogSender is disabled. Dependency will not be added to project '${project.path}'.")
                 }
 
-                if (findProperty(ToolingConfig.PROP_JDWP_INJECT)?.toString()?.toBoolean() == true) {
-                    val directory = findProperty(ToolingConfig.PROP_JDWP_LIBDIR)?.toString()
-                        ?: throw GradleException("Property '${ToolingConfig.PROP_JDWP_LIBDIR}' not set")
-
-                    logger.info(
-                        "Trying to inject JDWP library directory '{}' to project '{}'",
-                        directory,
-                        project.path
-                    )
-
-                    (extensions.getByType(AndroidComponentsExtension::class.java) as ApplicationAndroidComponentsExtension).apply {
-                        val debuggableVariants = mutableListOf<String>()
-                        beforeVariants { variantBuilder ->
-                            if (variantBuilder.debuggable) {
-                                debuggableVariants.add(variantBuilder.name)
-                            }
-                        }
-
-                        onVariants { variant ->
-                            if (variant.name in debuggableVariants) {
-                                logger.info(
-                                    "Injecting JDWP library directory '{}' to variant '{}' of project '{}'",
-                                    directory,
-                                    variant.name,
-                                    project.path
-                                )
-                                // add the directory to the list of static sources
-                                variant.sources.jniLibs?.addStaticSourceDirectory(directory)
-
-                                val debuggerConfigTask = tasks.registerTask(
-                                    "${variant.name}DebuggerConfig",
-                                    AppDebuggerConfigTask::class.java
-                                )
-
-                                debuggerConfigTask.configure { task ->
-                                    task.jdwpOptions.set(project.jdwpOptions())
-                                }
-
-                                variant.sources.java?.addGeneratedSourceDirectory(
-                                    debuggerConfigTask,
-                                    AppDebuggerConfigTask::javaOut
-                                )
-
-                                variant.artifacts.use(debuggerConfigTask)
-                                    .wiredWithFiles(
-                                        AppDebuggerConfigTask::manifestIn,
-                                        AppDebuggerConfigTask::manifestOut
-                                    )
-                                    .toTransform(SingleArtifact.MERGED_MANIFEST)
-                            }
-                        }
-                    }
+                if (isDebuggerEnabled) {
+                    logger.info("Applying JdwpConfig plugin to project '${project.path}'")
+                    pluginManager.apply(JdwpConfigPlugin::class.java)
                 }
             }
         }
